+++ src/gamsrv.py
@@ -1,58 +1,8 @@
 from collections import defaultdict
-
-
-class Node:
-    def __init__(self, value, priority):
-        self.value = value
-        self.priority = priority
-
-
-class PriorityQueue:
-    def __init__(self):
-        self.heap = []
-
-    def up(self, i):
-        while i != 0 and self.heap[i].priority > self.heap[(i - 1) // 2].priority:
-            self.heap[i], self.heap[(i - 1) // 2] = self.heap[(i - 1) // 2], self.heap[i]
-            i = (i - 1) // 2
-
-    def down(self, i):
-        n = len(self.heap)
-        while 2 * i + 1 < n:
-            max_child = 2 * i + 1
-            if max_child + 1 < n and self.heap[max_child].priority < self.heap[max_child + 1].priority:
-                max_child += 1
-            if self.heap[i].priority >= self.heap[max_child].priority:
-                break
-            self.heap[i], self.heap[max_child] = self.heap[max_child], self.heap[i]
-            i = max_child
-
-    def insert_element(self, obj):
-        self.heap.append(obj)
-        self.up(len(self.heap) - 1)
-
-    def delete_element(self, obj):
-        obj_index = None
-        n = len(self.heap)
-        for index in range(0, n):
-            if obj == self.heap[index]:
-                obj_index = index
-        self.heap[obj_index], self.heap[n - 1] = self.heap[n - 1], self.heap[obj_index]
-        self.heap.pop(n - 1)
-        self.down(obj_index)
-
-    def pop_element(self):
-        if not self.heap:
-            return None
-        obj = self.heap[0]
-        self.heap[0] = self.heap[-1]
-        self.heap.pop()
-        self.down(0)
-        return obj.value, obj.priority
+import src.heap_based_priority_queue as pqueue
 
 
 class GameServer:
-
     @staticmethod
     def read_input_data_from_file(input_file):
         if input_file is None:
@@ -88,8 +38,8 @@ class GameServer:
             reversed_graph[end_node].append((start_node, latency))
 
         def dijkstra(start_pos, input_graph):
-            pq = PriorityQueue()
-            priority_queue_start_node = Node(start_pos, 0)
+            pq = pqueue.PriorityQueue()
+            priority_queue_start_node = pqueue.Node(start_pos, 0)
             pq.insert_element(priority_queue_start_node)
             shortest_path = {}
 
@@ -98,7 +48,7 @@ class GameServer:
                 if end_pos not in shortest_path:
                     shortest_path[end_pos] = latency
                     for end_pos_i, latency_i in input_graph[end_pos]:
-                        priority_queue_add_node = Node(end_pos_i, latency + latency_i)
+                        priority_queue_add_node = pqueue.Node(end_pos_i, latency + latency_i)
                         pq.insert_element(priority_queue_add_node)
             return shortest_path
